Reorganize entire file structure and add react test files
