+++ src/MovieInfo/MovieInfo.js
