+++ src/Posters/Posters.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import './Posters.css';
-import Poster from './Poster.js'
-import Header from './Header.js'
+import Poster from '../Poster/Poster.js'
+import Header from '../Header/Header.js'
 
 const Posters = (props) => {
   const posters = props.posters
